add vueify
