+++ package.json
@@ -31,6 +31,10 @@
   },
   "devDependencies": {
     "autoprefixer": "^6.3.3",
+    "babel-core": "^6.7.4",
+    "babel-plugin-transform-runtime": "^6.6.0",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-runtime": "^5.8.38",
     "browser-sync": "^2.11.1",
     "browserify": "^13.0.0",
     "gulp": "^3.9.1",
@@ -43,6 +47,9 @@
     "gulp-sourcemaps": "^1.6.0",
     "imagemin-pngquant": "^4.2.2",
     "lost": "^6.7.2",
-    "vinyl-source-stream": "^1.1.0"
+    "vinyl-source-stream": "^1.1.0",
+    "vue-hot-reload-api": "^1.3.2",
+    "vueify": "^8.3.6",
+    "vueify-insert-css": "^1.0.0"
   }
 }
