+++ src/index.html
@@ -17,21 +17,12 @@
     ></menu>
 
     <div class="wrapper">
-      <router-view
-        :duo="$route.params.duo"
-      ></router-view>
+      <router-view></router-view>
 
       <button @click="goBack" v-show="canGoBack" class="button button--pagenav button--previous button--translateleft"><span class="wb-arrow-left"></span></button>
       <button @click="goForward" v-show="canGoForward" class="button button--pagenav button--next button--translateright"><span class="wb-arrow-right"></span></button>
     </div>
   </div>
-
-  @@include('./partials/menu-template.html')
-  @@include('./partials/book-template.html')
-  @@include('./partials/page-template.html')
-  @@include('./partials/modal-static-template.html')
-
-
   <script type="text/javascript" src="./js/bundle.js"></script>
 </body>
 </html>
