+++ src/js/components/book.vue
@@ -1,0 +1,34 @@
+<template>
+  <section class="book">
+    <page type="left" :id="left" :data="pages[left]"></page>
+    <page type="right" :id="right" :data="pages[right]"></page>
+  </section>
+</template>
+
+<script>
+  var pages = require('../pages.js');
+
+  export default {
+    name: 'book',
+
+    computed: {
+      left: function() {
+        var playdoh = this.$route.params.duo;
+        playdoh = playdoh.split('-');
+        return playdoh[0];
+      },
+
+      right: function() {
+        var playdoh = this.$route.params.duo;
+        playdoh = playdoh.split('-');
+        return playdoh[1];
+      }
+    },
+
+    data: function() {
+      return {
+        pages: pages()
+      }
+    }
+  }
+</script>
