--- src/partials/book-template.html
@@ -1,6 +1,0 @@
-<template id="book-template">
-  <section class="book">
-    <page type="left" :id="left" :data="pages[left]"></page>
-    <page type="right" :id="right" :data="pages[right]"></page>
-  </section>
-</template>
